Remove remaining indirect syscall usage in lang/go.
This removes the few remaining uses of indirect syscalls in Go's syscall
package, as well as those remaining in os/signal and syscall tests.

Unfortunately, this also means disabling the race detector on openbsd/amd64
since the race detector contains references to libc's syscall symbol. This
cannot currently be regenerated/updated, since LLVM TSAN went and removed
support for OpenBSD without discussing it with anyone.

Thanks to tb@ for testing this on amd64 and arm64 machines that do not have
indirect syscall support.

ok tb@