Import devel/p5-File-FcntlLock 0.22
File::FcntlLock is a module to do file locking in an object oriented
fashion using the fcntl(2) system call. This allows locks on parts of
a file as well as on the whole file and overcomes some known problems
with flock(2), on which Perls flock() function is based per default.

From MAINTAINER Renaud Allard
ok bluhm@