Remove references to the libc syscall stub from lang/go.
The syscall symbol will soon be removed from libc - remove references to it
from Go, so that it will continue to function. This effectively means that
syscall.Syscall* and syscall.RawSyscall* will stop working and return
ENOSYS. For the time being, given the amount of usage in the Go ecosystem,
reroute calls with SYS_IOCTl and SYS_SYSCTL to the appropriate libc stub,
so that they continue to work.

Thanks to tb@ for testing on systems with syscall removed from libc.