Switch LDEP from pcre to pcre2
The passenger build system checks for the pcre header, and will fail if
it is not present (even if pcre2 is installed).  However, the only place
pcre is used is for the embedded nginx, and that will pick pcre2 over
pcre if both are installed.  This switches the LDEP from pcre to pcre2
so that the embedded nginx will always use pcre2, and adds a BDEP on
pcre so that the passenger build system still works.

While here, comment out a file in the PLIST so updates using make
update-plist are easier.

Problem found by tb@ because dpb junked pcre2 while building passenger

OK tb@