step-cli: set HOME to WRKDIR to fix build in some setups, ok sthen
With the last update the code generating shell completions tries to write
into /root/.step/ when building ports with dpb in a proot (at least in my
configuration). EACCESS is fatal in while ENOENT isn't [1]. In more usual
setups, it would use the /nonexistent/.step via the homedir of _pbuild, so
it would work for those.

[1]: https://github.com/smallstep/cli-utils/blob/c3dd6955f2e4e285216beb6d4b3a452cdcaed5fe/step/context.go#L132-L138