Fix build of py-adblock after maturin update
This uses some patches stolen from arch linux.

tested/ok sthen who suggested a tweak to reduce PLIST churn

(As an aside, the local news papers are among the very few pages where
disabling the ad blocker actually reduces the amount of wanting to throw
things at the screen for me. Interestingly and quite unexpectedly that's
different in sthen's neck of the wood.)