Fix WANTLIB for Ruby 3.1 and 3.2, other minor changes
This sets PATCHORIG = .orig, because the build infrastructure handles
that better and it requires less work when updating plists.

This sets optflags="", so that the build infrastructure does not
override the default optimization flags.

This removes ncurses, readline, and yaml from WANTLIB, and adds yaml-0
to WANTLIB. For Ruby 3.1 and 3.2, ncurses and readline are added back.
Ruby 3.3 drops the readline extension, so those libraries are no longer
needed.

This drops the gdbm subpackage code from Makefile.inc, since it
is no longer used by any supported Ruby version.

OK tb@, gkoehler@