import ruby-faraday-net_http
This gem is a Faraday adapter for the Net::HTTP library. Faraday is an
HTTP client library that provides a common interface over many adapters.
Every adapter is defined into it's own gem. This gem defines the adapter
for Net::HTTP the HTTP library that's included into the standard library
of Ruby.

OK jeremy@