Switch lang/ghc to 9.6.4
The massive change is due to upstream switching to a new build system
called hadrian. It needs to be bootrapped separately and its sources
are kept as a bootstrap bundle.

Documentation is no longer deemed worth packaging due to space and
time overhead compared to its utility.

Most of the work happened at h2k23, so this is a belated thank you
to University of Coimbra and the portuguese usual suspects.

OK kili@