import nanosaur2-2.1.0
Nanosaur II: Hatchling is a continuation of the original Nanosaur
storyline, only this time you get to fly a pterodactyl who's loaded with
hi-tech weaponry.  In addition to the three level adventure mode where
you blast enemy dinos and save eggs, there are also six two-player
levels including racing, capture-the-flag, and battle.

This is a port of the original Mac version of the game to modern
operating systems, made under permission from Pangea Software, Inc.

several tweaks and ok gkoehler@