update nbconvert to 6.5.0
Now that we have all newly required ports in the tree we can update to a
newever version of nbconvert. This update will allow us to make forward
progress on getting jupyter_server imported which will allow us to further
update the jupyter stack.

Some notable changes in this update:
- patches can go away because this version now officially supports newer
  jinja2
- bleach 5.0 is now supported
- removed dep on legacy ipython_genutils port