import volatility3
Volatility is the world's most widely used framework for extracting
digital artifacts from volatile memory (RAM) samples. The extraction
techniques are performed completely independent of the system being
investigated but offer visibility into the runtime state of the system.
The framework is intended to introduce people to the techniques and
complexities associated with extracting digital artifacts from volatile
memory samples and provide a platform for further work into this
exciting area of research.

Symbol files are not part of the package. They have to be installed 
manually. See README for details.

valuable feedback and OK sthen@