Update to py3-fastecdsa 2.3.2
While the tests no longer reach out to the internet, we have a new
complication because for some reasons the tests look for the json
files in the wrong directory. Also, the test data didn't make it to
pypi. So grab the files from github and stupid patches to fix the paths.
Unsure if it's Python being too smart for its own good or python.port.mk.
Probably both.