update ipython to 8.22.1
In this update also remove the BDEP on www/py-mako which was added a
few years ago. That BDEP doesn't make a lot of sense. If we see new
build-time failures as result of this change, they should be looked at
again given the many updates in the ports tree since this BDEP was added.