remove Python 2 from early versions of automake
While automake before 1.11 needs Python 2, it may be not be needed for
anything in our tree except the automake regress tests. Switch to Python 3.

If anything breaks as a result of this change, we can explore modernizing
the Python in these older versions of automake.

(Part of a larger diff covering all versions of automake, but committing
in two pieces. This is the first piece.)

ok sthen@