import pipe-viewer-0.5.0
gtk-pipe-viewer is a lightweight application with a GTK interface (fork
of straw-viewer) for searching and playing videos from YouTube. This
fork parses the YouTube website directly and relies on the invidious
instances only as a fallback method.

from Stefan Moran, who is also taking MAINTAINER, thank you!
with some tweaks from me and kn, ok kn