Update to Suricata 7.0.4
Take co-maintainership.

Drop unnecessary AUTOMAKE version and SUBST_CMD processing of
configure and doc/userguide/Makefile.in.

Drop empty line patch of doc/userguide/Makefile.in.

Process suricata.yaml.in with SUBST_CMD, and patch it so all
@e_*@ variables are replaced.

As the pkg README recommends suricata-update for updates,
make that the default in the config file.

Fix the pkg README to set the path when calling suricata-update,
as the default path is not the desired path (it would be best to
fix the default path, but the current patches are insufficient).

Following the instructions in the README to setup suricata now
results in a working configuration without any configuration
file modification.

OK gonzalo@ (MAINTAINER)
OK sthen@