Import libinih-58
inih (INI Not Invented Here) is a simple .INI file parser written in C. It's
only a couple of pages of code, and it was designed to be small and simple, so
it's good for embedded systems. It's also more or less compatible with Python's
ConfigParser style of .INI files, including RFC 822-style multi-line syntax and
name: value entries.

ok robert@