Update to apertium-3.8.3. Update several other apertium-related ports.
This is a complex web of interdependencies, so it's better to update
these all at once. Also, there are newer versions of many of these
ports, but what's here builds and runs, so I am committing it as-is.

Two additional languages are marked broken, but are likely fixed
easily with further updates and/or backports.