Import Qt6 SignOn
The SignOn daemon is a D-Bus service which performs user authentication on
behalf of its clients. There are currently authentication plugins for OAuth 1.0
and 2.0, SASL, Digest-MD5, and plain username/password combination.

Background informations:
SignOn decided depending on the QMake version whether it is a Qt5 or Qt6
target. That makes it a freak show to build it in a port.

This is not a gift for for all PACKAGE MAINTAINER. qmake <3

OK landry@