Fix binary wheel file names on riscv64
Before the patch, pip install complained:
ERROR: jellyfish-1.0.3-cp310-cp310-openbsd_7_5_riscv.whl is not a supported wheel on this platform.

The arch is really named "riscv64", not "riscv" on OpenBSD. This may be
different from other systems but we're not going to support riscv32
systems.

Second commit only to add this to cvs log.  Thank you cvs(1).