import games/recoil-rts, ok solene@
DESCR:
Recoil is a battle tested open-source RTS engine, successor of Spring. It is
designed, in its basis, to be able to run the content of the game Total
Annihilation and deliver a similar, but improved, gaming experience. Games can
be intense and very large scaled, with fight of, literally, hundreds of units
and the mods allow very wide arrays of different strategies and tactics. Some
of the games powered by Recoil: Beyond All Reason, ZeroK, TA Prime and Metal
Factions.