Import kf6-kwayland-6.0.4
Comment:
Qt API to interact with the wayland-client and -server

Description:
KWayland provides two libraries:

- KWayland::Client
- KWayland::Server

As the names suggest they implement a Client respectively a Server API for the
Wayland protocol. The API is Qt-styled removing the needs to interact with a
for a Qt developer uncomfortable low-level C-API. For example the callback
mechanism from the Wayland API is replaced by signals; data types are adjusted
to be what a Qt developer expects, e.g. two arguments of int are represented by
a QPoint or a QSize.

Maintainer: Rafael Sadowski <rsadowski@openbsd.org>

OK landry@