p5-Crypt-OpenSSL-PKCS10: fix after X509V3_EXT_add_alias() removal
This isn't a complete fix but it is good enough for it to do what
henning needs it to do. For some reason this code was left in the
state it was in for ~20 years without any update. No idea why it
could not support things like basic constraints or subject/authority
info access using proper NIDs.

ok henning