MuJS is a lightweight Javascript interpreter designed for embedding in other software to extend them with scripting capabilities.
MuJS was designed with a focus on small size, correctness, and
simplicity. It is written in portable C and implements ECMAScript
as specified by ECMA-262. The interface for binding with native
code is designed to be as simple as possible to use, and is very
similar to Lua. There is no need to interact with byzantine C++
template mechanisms, or worry about marking and unmarking garbage
collection roots, or wrestle with obscure build systems.

ok sthen@ some time ago