databases: drop pg_statsinfo & pg_statsreporter
the former has frequently been in the way of postgresql upgrades, and
latest version parses /proc/{cpu,mem}info. It's time to give up, FreeBSD
dropped the port in 2018.

the latter depends on the former, and is an old php codebase which we
aren't fully sure works with recent php, so is in the way of dropping
old phps.