Drop GCC_PRINTFLIKE function attribute and fix build on sparc64
The GCC_PRINTFLIKE macro was slightly changed in the last ncurses update
and now breaks the way it is used by pftop.  base-gcc doesn't
recognize the attribute syntax, while ports-gcc instead complains:

error: attributes should be specified before the declarator in
a function definition

Moving the attribute before the declarator fixes the build. Just drop
the attribute as it's not useful here (the declarations of both
functions are already annotated elsewhere) and they aren't present
either in usr.bin/systat/engine.c.