import devel/rust-bindgen 0.69.4
ok robert@

bindgen automatically generates Rust FFI bindings to C (and some C++)
libraries.

bindgen cli tool takes the path to an input C or C++ header file, and
optionally an output file path for the generated bindings. If the
output file path is not supplied, the bindings are printed to stdout.

If we wanted to generated Rust FFI bindings from a C header named
input.h and put them in the bindings.rs file, we would invoke bindgen
like this:

$ bindgen input.h -o bindings.rs

For more details, pass the --help flag:
$ bindgen --help