import games/indierunner ok bentley@ testing and feedback by fabien@
Successor of and expansion on games/fnaify, with the goal of
facilitating launch of many game engines that may require complex file
system and runtime configuration to run on OpenBSD. This is an early
version that may contain bugs; package will allow for broader testing
to iron out issues prior to an eventual 1.0 release given the large
library of hundreds of games.

DESCR:
IndieRunner provides a universal CLI launcher for a variety of independent
games built with frameworks like FNA, XNA, LWJGL, LibGDX, HashLink, GZDoom,
ScummVM, Love2D. It selects a runtime, configures it as needed, and takes care
of outdated or incompatible bundled libraries.