Enable VA-API in graphics/ffmpeg
This is another small step towards making VA-API usable in ports tree.
Keep in mind that this does not mean that VA-API will work everywhere.

Especially for the browsers there is still a lot of work to do. mpv(1) and some
others are happy with it though!

Special thanks to José Maldonado and Lucas Gabriel Vuotto for testing and
giving feedback. And of course all ports/xenocara hackers!

Feedback and OK Brad (Maintainer), tb@