partially revert some parts of previous ghostscript commit;
- disable gxps and gpdl; both are produced from the xps source dir,
which produces a libgxps shared library that conflicts with the existing
(different) shared library in graphics/libgxps.

it might be possible to reenable parts of this (maybe just removing
gxps from the packages), but care will also be needed to make sure that
it doesn't pick up an existing /usr/local/lib/libgxps.so.* during the
build.

- @comment the shared library symlinks, OpenBSD does not use these.

- update flavours list in DESCR while there