update to gotosocial-0.16.0 and unbreak, ok Hukadan (maintainer)
- fails with go 1.23rc2 due to https://go.dev/doc/go1.23#linker:
it seems codeberg.org/gruf/go-fastcopy is not part of the "large
open-source code corpus" using //go:linkname which remains supported -
but we have a workaround (commented-out in Makefile).

- upstream uses a bogus version number for modernc.org/sqlite as a
placeholder for a version which gets changed by a "replace" line
in go.mod to use a fork; this results in modgo-gen-modules trying
to fetch a nonexistent v0.0.0-00010101000000-000000000000.mod
which isn't really used - remove this MDOGO_MODFILES entry to.work
around. (It seems modgo-gen-modules may fetch more .mod files than
are really needed anyway, even outside this special situation).