import devel/py-kaitaistruct, from Kirill Korinsky, ok daniel@
This library implements Kaitai Struct API for Python.

Kaitai Struct is a declarative language used for describe various binary
data structures, laid out in files or in memory: i.e. binary file
formats, network stream packet formats, etc.

It is similar to Python's Construct 2.10 but it is language-agnostic.
The format description is done in YAML-based .ksy format, which then can
be compiled into a wide range of target languages.