Unbreak building of sysutils/rustic
rustic uses env::current_exe() to get the full file system path of the
executable, which is not supported on OpenBSD. To work around this the
path has been hard coded following a couple of other ports (e.g.,
devel/cargo-audit and lang/deno). This hard coding breaks building as
rustic itself is used in the post-build phase to generate
bash-/fish-/zsh-completions. In that phase the binary is not yet copied
to the hard coded path.

Issue is not seen when building on a system with rustic installed.

REVISION has been bumped, just to be sure.

Prodded by naddy@ and ajacoutot@