import sysutils/fswatch 1.17.1
ok sthen@

HOMEPAGE: https://emcrisostomo.github.io/fswatch/

fswatch monitor files and directory trees for changes.

Main features are:

- Support for many OS-specific APIs such as kevent, inotify, and FSEvents.
- Recursive directory monitoring.
- Path filtering using including and excluding regular expressions.
- Customizable record format.
- Support for periodic idle events.

Limitations:

The kqueue monitor, available on any *BSD system featuring kqueue, requires a
file descriptor to be opened for every file being watched. As a result, this
monitor scales badly with the number of files being observed, and may begin to
misbehave as soon as the fswatch process runs out of file descriptors. In this
case, fswatch dumps one error on standard error for every file that cannot be
opened.