import games/trackballs 1.3.4.
Trackballs is a simple game similar to the classical game Marble Madness
on the Amiga in the 80's. By steering a marble ball through a labyrinth
filled with vicious hammers, pools of acid and other obstacles the player
collects points. When the ball reaches the destination it continues
at the next, more difficult level--unless the time runs out.

based on an original submission 10 years ago by bentley@,
with tweaks and ok bentley@