update Godot to 3.6 and add support for GDNative modules via location ${LOCALBASE}/lib/godot/gdnative. This allows for other ports to build GDNative .so libraries and place them in this location so that they can be used at runtime.
The mechanism extends a series of FileAccess checks for other locations
of GDNative libraries from upstream. Discussed with op@ - while this is
a TOCTOU issue, at this point benefit of staying with upstream
practices outweighs this for upstreaming the diffs eventually. Plan is
to raise the TOCTOU issue separately with upstream.

Tested with multiple consumers and editor including godot-sharp.
(Re-)Add me to MAINTAINER in addition to op@
ok op@