import x11/elementary-xfce-icon-theme 0.20.
This is an icon-theme maintained with Xfce in mind, but it should work on
other desktops like Gnome3 as well.

It is a fork of the upstream elementary project. The reason for forking
was that the project decided to focus on its own desktop environment and
dropped a ton of (ugly, but necessary) symlinks. This icon-theme is
supposed to keep everything working, but gets updates from upstream
occasionally.

ok rsadowski@