Simplify getgrouplist(3) patch
Now that our implementation can be used to query the total number of
groups, stop avoiding the original code in get_user_groups().  Sadly
there's still an issue with the Linux implementation being incompatible
(there the function returns the number of groups found, not 0 on success).

Pointed out by ajacoutot@