Don't depend on gnome-initial-setup; it's going to the CVS Attic soon. Initial Setup is used to configure / create new users.
Currently, only if there is no user in the system, GDM will start Initial
Setup (as the _gnome-initial-setup user) and configure one. This may change in
the future and run for existing users that haven't connected to GNOME yet.

It's integration with PAM, chpasswd, /dev/dri/card* ownership etc. makes it
difficult running on OpenBSD.