www/nextcloud: assorted monthly updates, and retire 26
- add a MESSAGE file to all non-latest branches strongly nudging users
  to update to the latest upstream branch (as of now, 30, per
https://github.com/nextcloud/server/wiki/Maintenance-and-Release-Schedule)
- unlink 26, was EOL'ed upstream in march, force-update users to 27 via
  @pkgpath
- update 28 to 28.0.12 (EOL 2024-12)
- update 29 to 29.0.9  (EOL 2025-04)
- update 30 to 30.0.2  (EOL 2025-09)

users of nextcloud on -current are urged to update to 30, since by 2025
we'll retire 27/28/29 so that 7.7 only ships with 30 (and 31 by then..)
and we don't get a boatload of unmaintained versions. The last versions
of each branch will still be available in 7.6-stable packages.

pkg_add -r nextcloud%${BRANCH} is your friend :)

ok sthen@
gonzalo@ (MAINTAINER) agreed on the direction last we talked about it