Add vulkan as default build dependency for Qt6
Because of recurring Qt6/CMake build error messages that vulkan is missing. We
have decided to include this as a build dependency in the Qt6 module. Ports
that needs Vulkan as runtime should include it as usual.

While here cleanup double build depends entries

Feedback sthen@, tested in a full bulk build