Import devel/libvolk 3.1.2.
VOLK is the Vector-Optimized Library of Kernels. It is a library that
contains kernels of hand-written SIMD code for different mathematical
operations. Since each SIMD architecture can be very different and no
compiler has yet come along to handle vectorization properly or highly
efficiently, VOLK approaches the problem differently.

For each architecture or platform that a developer wishes to vectorize
for, a new proto-kernel is added to VOLK. At runtime, VOLK will select
the correct proto-kernel. In this way, the users of VOLK call a kernel
for performing the operation that is platform/architecture agnostic.
This allows us to write portable SIMD code.

This was spun off as a distinct project from comms/gnuradio in version
3.10.

thanks thfr@ for tweaking graphics/volk to properly deconflict both
ports.
ok tb@