Stop installing libbfd-related libs, headers etc from devel/gdb
This is something that would be more appropriate in devel/binutils, but
since nothing currently uses libbfd in the ports tree, just stop
shipping it in devel/gdb.

The initial reason was that the installed /usr/local/include/ansidecl.h
resulted in spurious conflicts with other copies in the ports tree,
spotted by at least naddy@ and tb@ since the update to gdb-13.2.

ok tb@ sthen@ pascal@ (maintainer)