import llvm-openmp 16.0.6 OK asou@
The version number is in sync with our base clang.  Our ld.so still
needs some fixes to make the OpenMP test suite pass.  Basically the
dtor destructor callbacks are not invoked right now.  And the order
of destructor calls seems to be different on GNU/Linux.

Comment:
LLVM OpenMP libraries

Description:
The OpenMP API supports multi-platform shared-memory parallel
programming in C/C++ and Fortran. The OpenMP API defines a portable,
scalable model with a simple and flexible interface for developing
parallel applications on platforms from the desktop to the
supercomputer.