carry out a few more checks on MODPY_PYBUILD settings, which identify some common mistakes (in particular, missing changes to build system done with an update).
not fatal because sometimes there may be a reason for it not matching up,
but it's not resulting in many false warnings.

defer printing this until fake-install where it's more likely to be seen
(otherwise it's likely to be hidden by a wall of scrolling messages with
some build systems).