Update to py3-fastecdsa 3.0.0
Rename the new benchmark script to fastecdsa-benchmark and relax setuptools
version requirement.

This port always seems to require some stupid hacks to get working tests.
This time it ships with an empty __init__.py that breaks the hell out of
everything. Disable py-test-cov nonsense. This doesn't seem to work here.
It isn't particularly interesting for a package anyway.