Fix build issues for lang/php/*
sthen@ reported a buid issue with php after a recent xxhash update.

The build for php-8.2 and php-8.3 was broken due to an oversight in
handling a recent update to xxhash. Specifically, members of a struct in
a public header were renamed, causing issues as xxhash is used as a
header-only library.

php-8.4 inadvertently built successfully as it unintentonally used a
packaged header file instead of the one provided by the xxhash port.

This commit resolves the build failures for php-8.2 and php-8.3 and
ensures that php-8.4 uses the correct header file from xxhash from
ports.

OK tb@, sthen@