import ports/astro/py-astropy-iers-data, tweak/ok aja@ (+ I fleshed out DESCR a bit after sending for review).
astropy-iers-data provides access to the tables provided by the
International Earth Rotation and Reference Systems (IERS) service,
in particular the Earth Orientation data allowing interpolation of
published UT1-UTC and polar motion values for given times.

It is not meant to be used directly, but by the core astropy package.

The UT1-UTC values are used in Time and Dates (astropy.time) to provide
UT1 values, and the polar motions are used in astropy.coordinates to
determine Earth orientation for celestial-to-terrestrial coordinate
transformations.