jujutsu: switch to vendored libgit2
libgit2 is used for clone, fetch, push and broke with the 1.9.0 update:
Error: invalid version 0 on git_proxy_options; class=Invalid (3)
which is an API break not mentioned in the libgit2 release notes. Most
other git-rs/libgit2-sys consumers likely don't do networking and also
don't seem to be using blame or .gitconfig things and can be fixed if
users complain.