Retire OpenSSL 1.1 and 3.2, set @pkgpath for 3.3; help/ok sthen
Going forward, there will only be two OpenSSL releases in the ports tree:
the latest stable version for testing (currently 3.4) and the penultimate
stable version (currently 3.3) for ports to consume.

Soon after the release of 3.5.0, consumers will switch to 3.4 and 3.3 will
be removed, and so on.  The idea is that this way, 3.4 will have had some
time to mature and the latest round of regressions should have been found
and addressed in 3.4.1 or 3.4.2 and it should be safe enough to switch.