import games/exult; ok volker@
DESCR:
As computers evolve, we risk losing classic games from the not-so-distant past.
One of these, Ultima VII, a two part RPG from the early 1990's, still has a huge
following, and many consider it to be one of the most immersive games ever
produced. Exult is a project to recreate Ultima VII for modern operating
systems, using the game's original plot, data, and graphics files.