Move ports back to using --forward in PATCH_ARGS and PATCH_DIST_ARGS. this was used earlier, but lost when changing to PATCH_DEBUG by default.
Without this, if a diff appears to be reversed, patch asks whether to
assume that you used -R, with a default 'y'. If there's no tty, as is the
case with a bulk build, the default is accepted, resulting in problems
(potentially nasty in the case of a security fix which was applied via
a patch and later committed upstream) if 'cvs rm' was missed.

Sync option order between PATCH_ARGS and PATCH_DIST_ARGS while there.

Grepping ports logs doesn't show any current ports where this will cause
unexpected changes in behaviour.

ok tb