Import games/stratagus, free cross-platform real-time strategy gaming engine.
With help from and OK thfr@ and rsadowski@

Stratagus is a free cross-platform real-time strategy gaming engine. It
includes support for playing over the internet/LAN, or playing a
computer opponent. The engine is configurable and can be used to create
games with a wide-range of features specific to your needs. Stratagus is
written in C++.