Import games/wargus, Warcraft 2 Mod
With help from and OK thfr@ and rsadowski@

Wargus is a Warcraft 2 Mod that allows you to play Warcraft 2 with the Stratagus
engine. The game looks and sounds almost exactly like Warcraft 2. The gameplay
is very close to Warcraft 2's gameplay, with slight enhancements towards
Starcraft and other more modern RTS games.

While this package and the game engine are open source, one still needs
to provide the original game assets from Warcraft II, which is not open
source. You will need to purchase the original game to use this package.