Import games/war1gus, Warcraft Orcs and Humans Mod
With help from and OK thfr@ and rsadowski@

War1gus is a re-implementation of "Warcraft: Orcs & Humans" that allows you to
play Warcraft with the Stratagus engine. The game uses graphics and sounds from
the original Warcraft, but improves the gameplay mechanisms with many modern
conveniences that the Stratagus engine allows, such as modern mouse controls,
named groups, larger group selection, more player factions in multiplayer games,
a map editor, and multiple towns.

While this package and the game engine are open source, one still needs
to provide the original game assets from Warcraft Orcs and Humans, which
is not open source. You will need to purchase the original game to use
this package.